+++ app/controllers/admin/team_members_controller.rb
@@ -1,7 +1,15 @@
 class Admin::TeamMembersController < ApplicationController
-  before_action :authenticate_user!
-  before_action :check_owner
-  before_action :set_admin_team_member, only: [:show, :edit, :update, :destroy]
+  before_action :authenticate_user!,    except: [:photo]
+  before_action :check_owner,           except: [:photo]
+  before_action :set_admin_team_member, only: [:show, :edit, :update, :destroy, :photo]
+
+  def photo
+    if @admin_team_member.photo.size > 0
+      send_file File.join(CarrierWave::Uploader::Base.root, @admin_team_member.photo.url)
+    else
+      head :ok
+    end
+  end
 
   # GET /admin/team_members
   # GET /admin/team_members.json