+++ app/helpers/highlights_helper.rb
@@ -4,4 +4,12 @@ module HighlightsHelper
     ['http', 'https', 'ftp', 'mailto'].include?(URI.parse(url).scheme) rescue false
   end
 
+  def highlight_photo_url(highlight)
+    if highlight.photo.present?
+      photo_highlight_path(highlight)
+    else
+      nil
+    end
+  end
+
 end