+++ app/helpers/users_helper.rb
@@ -1,2 +1,11 @@
 module UsersHelper
+
+  def user_photo_url(user)
+    if user.photo.present?
+      photo_user_path(user)
+    else
+      'default-user.jpg'
+    end
+  end
+
 end