+++ app/models/admin/partner.rb
@@ -17,11 +17,4 @@ class Admin::Partner < ActiveRecord::Base
 
   # Named scopes
   # Methods
-  def photo_url
-    if photo.present?
-      photo.url
-    else
-      'default-user.jpg'
-    end
-  end
 end