+++ app/uploaders/photo_uploader.rb
@@ -10,7 +10,7 @@ class PhotoUploader < CarrierWave::Uploader::Base
   # Override the directory where uploaded files will be stored.
   # This is a sensible default for uploaders that are meant to be mounted:
   def store_dir
-    File.join(model.class.to_s.underscore, mounted_as.to_s, model.id.to_s)
+    File.join(model.class.to_s.underscore.pluralize, model.id.to_s, mounted_as.to_s)
   end
 
   # Provide a default URL as a default if there hasn't been a file uploaded: